Add getUserMessage() function to response object
I am proposing this because some gateways return messages with supplementary data - e.g. Cybersource returns a message indicating invalid data and in a separate field the relevant field names. Also, some processors return reponses with 2 different levels of details e.g 'your transaction was not processed' to the user and 'arrest this person now' to the admin